requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-appconf==1.0.5
django-axes==5.35.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-imagekit==4.1.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-organizations==2.0.2
django-phonenumber-field==6.3.0
django-polymorphic==3.1.0
django-rest-auth==0.9.5
django-rest-polymorphic==0.1.9
django-simple-history==3.1.1
django-timezone-field==5.0
djangoql==0.15.4
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
flower==1.1.0
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.51
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
